+++ src/index.ts
@@ -1,5 +1,5 @@
 import { ApolloServer } from '@apollo/server';
-import { startStandaloneServer } from '@apollo/server/standalone';
+import { startStandaloneServer, } from '@apollo/server/standalone';
 import { addMocksToSchema } from '@graphql-tools/mock';
 import { makeExecutableSchema } from '@graphql-tools/schema';
 
@@ -11,6 +11,11 @@ import { typeDefs } from './schema';
 import { resolvers } from './resolver';
 import { mocks } from './mocks';
 
+// TODO: We likely want to switch this up to validate the token and return the User in the JWT
+function getTokenFromRequest(request) {
+  return request.headers?.authentication || '';
+}
+
 async function startApolloServer() {
   const server = new ApolloServer({
     schema: addMocksToSchema({
@@ -18,17 +23,24 @@ async function startApolloServer() {
       mocks,
       preserveResolvers: true,
     }),
+    // Mitigation for an issue that causes Apollo server v4 to return a 200 when a query
+    // includes invalid variables.
+    //    See: https://www.apollographql.com/docs/apollo-server/migration/#known-regressions
+    status400ForVariableCoercionErrors: true
   });
 
+  const { cache } = server;
   const { url } = await startStandaloneServer(server, {
-    context: async ({ req }) => ({
-      dataSources: {
-        // token: req.headers?.authentication
-        // cache: ???
-        dmphubAPIDataSource: await new DMPHubAPI({}),
-        sqlDataSource: await new MysqlDataSource({ config: mysqlConfig }),
-      },
-    }),
+    context: async ({ req }) => {
+      const token = getTokenFromRequest(req);
+
+      return {
+        dataSources: {
+          dmphubAPIDataSource: await new DMPHubAPI({ cache, token }),
+          sqlDataSource: await new MysqlDataSource({ config: mysqlConfig }),
+        },
+      }
+    },
   });
   console.log(`
     🚀  Server is running!