+++ src/mocks.ts
@@ -1,19 +1,18 @@
 import casual from 'casual';
-import { DmspId } from './schemas/scalars/dmspId';
-import { Orcid } from './schemas/scalars/orcid';
-import { Ror } from './schemas/scalars/ror';
+import { DMSP_BASE_URL, validateDmspId } from './resolvers/scalars/dmspId';
+import { validateOrcid } from './resolvers/scalars/orcid';
+import { Ror } from './resolvers/scalars/ror';
 import { mock as userMock } from './mocks/user';
 
 // Mock resolvers for our custom Scalars
 function mockOrcid() {
-  const id = casual.card_number().toString().match(/[0-9]{4}/g).join('-');
-  return new Orcid(`${Orcid.baseURL}${id}`);
+  return validateOrcid(casual.card_number().toString().match(/[0-9]{4}/g).join('-'));
 }
 function mockRor() {
   return new Ror(`${Ror.baseURL}${casual.rgb_hex.replace('#', '')}`);
 }
 function mockDmspId() {
-  return new DmspId(`${DmspId.baseURL}${casual.rgb_hex.replace('#', '').toUpperCase()}`);
+  return validateDmspId(`${DMSP_BASE_URL}${casual.rgb_hex.replace('#', '').toUpperCase()}`);
 }
 
 export const scalarMocks = {