+++ src/resolver.ts
@@ -1,7 +1,7 @@
 import { mergeResolvers } from '@graphql-tools/merge';
 import { IResolvers } from '@graphql-tools/utils';
 
-import { orcidScalar } from './schemas/scalars/orcid';
+import { orcidScalar } from './resolvers/scalars/orcid';
 
 import { resolvers as contributorRoleResolvers } from './resolvers/contributorRole';
 import { resolvers as dmspResolvers } from './resolvers/dmsp';