+++ src/resolvers/scalars/ror.ts
@@ -2,7 +2,8 @@ import { GraphQLScalarType, Kind } from 'graphql';
 
 // An organization ROR ID
 export class Ror {
-  value: string;
+  identifier: string;
+  type: string;
 
   static baseURL: string = 'https://ror.org/';
   static idRegex: RegExp = /^[0-9a-zA-Z]+$/;
@@ -11,7 +12,8 @@ export class Ror {
   static validFormats: string = '"https://ror.org/abcd1234"';
 
   constructor(val: string) {
-    this.value = this.validate(val);
+    this.identifier = this.validate(val);
+    this.type = 'ror';
   };
 
   validate(val) {
@@ -29,14 +31,14 @@ export const rorScalar = new GraphQLScalarType({
 
   serialize(ror) {
     if (ror instanceof Ror) {
-      return ror.value; // Convert outgoing ROR to a String for JSON
+      return { type: ror.type, identifier: ror.identifier};
     }
     throw Error('GraphQL Ror Scalar serializer expected an `Ror` object');
   },
 
   parseValue(value) {
     if (typeof value === 'string') {
-      return new Ror(value); // Convert incoming string to ROR
+      return new Ror(value);
     }
     throw new Error(`GraphQL Ror Scalar expected a string in the ${Ror.validFormats} formats`);
   },
@@ -45,6 +47,8 @@ export const rorScalar = new GraphQLScalarType({
     if (ast.kind === Kind.STRING) {
       // Convert hard-coded AST string to string and then to Ror
       return new Ror(ast.value.toString());
+    } else if (ast.kind === Kind.OBJECT) {
+      return new Ror(ast.fields['identifier']);
     }
     // Invalid hard-coded value (not a string)
     return null;