+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Initial resolver to fetch a single DMSP from the DMPHub API
 - Initial DMPHub API data source
 - Initial MySQL data source
+- Custom GraphQL scalars for ROR, ORCID and DMSP IDs
 - Mechanism for Apollo to use mocks when a resolver has not yet been implemented
 - Mocks for User
 - Data migration mechanism `./data-migrations/process.sh`